// sfr-serverless/common/resources/notifications/types.ts
export enum NotificationKeyType {
  SentFriendRequest = "sent_friend_request",
  AcceptedFriendRequest = "accepted_friend_request",
  MatchPlay = "match_play",
  MatchPlayAccepted = "match_play_accepted",
  MatchPlayDeclined = "match_play_declined",
  MatchPlayRequested = "match_play_requested",
  MatchPlayRequestAccepted = "match_play_request_accepted",
  MatchPlayRequestDeclined = "match_play_request_declined",
  UploadScoreCard = "match_upload_scorecard",
  PlayerScoreCardUploaded = "player_scorecard_uploaded",
  TeamScoreCardUploaded = "team_scorecard_uploaded",
  MatchResultDeclared = "match_result_declared",
  MatchStarted = "match_started",
  MatchApproved = "match_approved",
  MatchFinished = "match_finished"
}